src/requirements.txt
connexion==2.14.1
environs==10.3.0
Flask==2.2.2
Flask-Cors==4.0.0
Flask-Migrate==3.1.0
PyJWT==2.8.0
PyMySQL==1.1.0
SQLAlchemy==2.0.25
SQLAlchemy-serializer==1.4.1
swagger-ui-bundle==0.0.9
cryptography==42.0.2